+++ source/core/parameters.cc
@@ -824,6 +824,13 @@ namespace Parameters
         "stabilization term on the VOF phase fraction <true|false>");
 
       prm.declare_entry(
+        "vof dcdd diffusion factor",
+        "0.5",
+        Patterns::Double(),
+        "Diffusion factor scaling the DCDD stabilization term in the VOF "
+        "equation");
+
+      prm.declare_entry(
         "pressure scaling factor",
         "1",
         Patterns::Double(),
@@ -872,6 +879,7 @@ namespace Parameters
       heat_transfer_dcdd_stabilization =
         prm.get_bool("heat transfer dcdd stabilization");
       vof_dcdd_stabilization = prm.get_bool("vof dcdd stabilization");
+      dcdd_diffusion_coeff   = prm.get_double("vof dcdd diffusion factor");
 
       pressure_scaling_factor = prm.get_double("pressure scaling factor");
     }